// src/types.ts
export interface Ticket {
  id: string
  company_name: string
  company_logo: string
  price_rub: number
  price_usd: number
  price_eur: number
  start_time: string
  end_time: string
  start_date: string
  end_date: string
  city_from: string
  city_to: string
  airport_from: string
  airport_to: string
  transfer: string
}